pyproject.toml
[tool.poetry]
name = "deep_image_blending"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^9.3.0"
scikit-image = "^0.19.3"
torch = "^1.13.1"
torchvision = "^0.14.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
